requirements.txt
alembic==1.9.4
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.1
Flask==2.2.3
Flask-Admin==1.6.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
platformdirs==3.1.1
pluggy==1.0.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.2
pytest-env==0.8.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
virtualenv==20.21.0
Werkzeug==2.2.3
WTForms==3.0.1
